+++ lib/src/usage_impl.dart
@@ -60,8 +60,11 @@ class AnalyticsImpl implements Analytics {
   static const String _defaultAnalyticsUrl =
       'https://www.google-analytics.com/collect';
 
+  @override
   final String trackingId;
+  @override
   final String applicationName;
+  @override
   final String applicationVersion;
 
   final PersistentProperties properties;
@@ -72,6 +75,7 @@ class AnalyticsImpl implements Analytics {
 
   final List<Future> _futures = [];
 
+  @override
   AnalyticsOpt analyticsOpt = AnalyticsOpt.optOut;
 
   String _url;
@@ -91,6 +95,7 @@ class AnalyticsImpl implements Analytics {
 
   bool _firstRun;
 
+  @override
   bool get firstRun {
     if (_firstRun == null) {
       _firstRun = properties['firstRun'] == null;
@@ -103,9 +108,7 @@ class AnalyticsImpl implements Analytics {
     return _firstRun;
   }
 
-  /**
-   * Will analytics data be sent?
-   */
+  @override
   bool get enabled {
     bool optIn = analyticsOpt == AnalyticsOpt.optIn;
     return optIn
@@ -113,13 +116,12 @@ class AnalyticsImpl implements Analytics {
         : properties['enabled'] != false;
   }
 
-  /**
-   * Enable or disable sending of analytics data.
-   */
+  @override
   set enabled(bool value) {
     properties['enabled'] = value;
   }
 
+  @override
   Future sendScreenView(String viewName, {Map<String, String> parameters}) {
     Map<String, dynamic> args = {'cd': viewName};
     if (parameters != null) {
@@ -128,8 +130,9 @@ class AnalyticsImpl implements Analytics {
     return _sendPayload('screenview', args);
   }
 
-  Future sendEvent(String category, String action, {String label, int value,
-      Map<String, String> parameters}) {
+  @override
+  Future sendEvent(String category, String action,
+      {String label, int value, Map<String, String> parameters}) {
     Map<String, dynamic> args = {'ec': category, 'ea': action};
     if (label != null) args['el'] = label;
     if (value != null) args['ev'] = value;
@@ -139,11 +142,13 @@ class AnalyticsImpl implements Analytics {
     return _sendPayload('event', args);
   }
 
+  @override
   Future sendSocial(String network, String action, String target) {
     Map<String, dynamic> args = {'sn': network, 'sa': action, 'st': target};
     return _sendPayload('social', args);
   }
 
+  @override
   Future sendTiming(String variableName, int time,
       {String category, String label}) {
     Map<String, dynamic> args = {'utv': variableName, 'utt': time};
@@ -152,12 +157,14 @@ class AnalyticsImpl implements Analytics {
     return _sendPayload('timing', args);
   }
 
+  @override
   AnalyticsTimer startTimer(String variableName,
       {String category, String label}) {
     return new AnalyticsTimer(this, variableName,
         category: category, label: label);
   }
 
+  @override
   Future sendException(String description, {bool fatal}) {
     // We trim exceptions to a max length; google analytics will apply it's own
     // truncation, likely around 150 chars or so.
@@ -181,8 +188,10 @@ class AnalyticsImpl implements Analytics {
     return _sendPayload('exception', args);
   }
 
+  @override
   dynamic getSessionValue(String param) => _variableMap[param];
 
+  @override
   void setSessionValue(String param, dynamic value) {
     if (value == null) {
       _variableMap.remove(param);
@@ -191,8 +200,10 @@ class AnalyticsImpl implements Analytics {
     }
   }
 
+  @override
   Stream<Map<String, dynamic>> get onSend => _sendController.stream;
 
+  @override
   Future waitForLastPing({Duration timeout}) {
     Future f = Future.wait(_futures).catchError((e) => null);